requirements.txt
beautifulsoup4
lxml
discord.py
python-dotenv
requests